pyproject.toml
[tool.poetry]
name = "fast-test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
pydantic-settings = "^2.2.1"
itsdangerous = "^2.2.0"
sqlalchemy = "^2.0.29"
pymysql = "^1.1.0"
starlette-async-jinja = "1.3.5"
aiopath = "0.6.13"
msgpack = "^1.0.8"
msgspec = "^0.18.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
